backend/pyproject.toml
[tool.poetry]
name = "django-api-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GNU"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
